// src/generated/api/identity/schema.ts
export interface components {
    schemas: {
        AuthTokensResponse: {
            accessToken: string;
            /** Format: date-time */
            accessTokenExpiresAt: string;
            isEmailVerified: boolean;
            isMobileVerified: boolean;
            refreshToken: string;
            /** Format: date-time */
            refreshTokenExpiresAt: string;
            requiresTotp: boolean;
            sessionId: string;
            userId: string;
        };
        CurrentUserResponse: {
            /** Format: email */
            email: string;
            isEmailVerified: boolean;
            isMobileVerified: boolean;
            mobileNumber: string;
            nationalCode: string;
            roles: string[];
            status: string;
            totpEnabled: boolean;
            userId: string;
        };
        OtpChallengeResponse: {
            challengeId: string;
            channel: string;
            destination: string;
            /** Format: date-time */
            expiresAt: string;
        };
        ProblemFieldError: {
            code: string;
            message: string;
        };
        ProblemResponse: {
            code: string;
            errors?: components["schemas"]["ProblemFieldError"][];
            message: string;
        };
        RefreshTokenRequest: {
            deviceId: string;
            deviceName: string;
            refreshToken: string;
        };
        SendOtpRequest: {
            channel: string;
            purpose: string;
        };
        SessionResponse: {
            /** Format: date-time */
            createdAt: string;
            deviceId: string;
            deviceName: string;
            /** Format: date-time */
            lastSeenAt: string;
            /** Format: date-time */
            refreshTokenExpiresAt: string;
            /** Format: date-time */
            revokedAt?: string | null;
            sessionId: string;
            status: string;
        };
        SignInRequest: {
            deviceId: string;
            deviceName: string;
            /** Format: email */
            email: string;
            password: string;
        };
        SignUpRequest: {
            deviceId: string;
            deviceName: string;
            /** Format: email */
            email: string;
            mobileNumber: string;
            nationalCode: string;
            password: string;
        };
        TotpEnrollmentResponse: {
            /** Format: date-time */
            createdAt: string;
            manualEntryKey: string;
            secret: string;
        };
        VerifyOtpRequest: {
            challengeId: string;
            code: string;
        };
        VerifyTotpRequest: {
            code: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}